% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootdht.R
\name{bootdht}
\alias{bootdht}
\title{Bootstrap uncertainty estimation for distance sampling models}
\usage{
bootdht(model, flatfile, resample_strata = FALSE, resample_obs = FALSE,
  resample_transects = TRUE, nboot = 100, summary_fun,
  convert.units = 1, select_adjustments = FALSE)
}
\arguments{
\item{model}{a model fitted by \code{\link{ds}} or a list of models}

\item{flatfile}{Data provided in the flatfile format. See \code{\link{flatfile}} for details.}

\item{resample_strata}{should resampling happen at the stratum (\code{Region.Label}) level? (Default \code{FALSE})}

\item{resample_obs}{should resampling happen at the observation (\code{object}) level? (Default \code{FALSE})}

\item{resample_transects}{should resampling happen at the transect (\code{Sample.Label}) level? (Default \code{TRUE})}

\item{nboot}{number of bootstrap replicates}

\item{summary_fun}{function that is used to obtain summary statistics from the bootstrap, see Details below.}

\item{convert.units}{conversion between units for abundance estimation, see "Units", below. (Defaults to 1, implying all of the units are "correct" already.)}

\item{select_adjustments}{select the number of adjustments in each bootstrap, when \code{FALSE} the exact detection function specified in \code{model} is fitted to each replicate. Note that for this to work \code{model} must have been fitted with \code{adjustment!=NULL}.}
}
\description{
Performs a bootstrap for simple distance sampling models using the same data structures as \code{\link[mrds]{dht}}.
}
\section{Model selection}{

Model selection can be performed on a per-replicate basis within the bootstrap. This has three variations:
\enumerate{
   \item when \code{select_adjustments} is \code{TRUE} then adjustment terms are selected by AIC within each bootstrap replicate (provided that \code{model} had the \code{order} and \code{adjustment} options set to non-\code{NULL}.
   \item if \code{model} is a list of fitted detection functions, each of these is fitted to each replicate and results generated from the one with the lowest AIC.
   \item if both 
   \item when \code{select_adjustments} is \code{TRUE} and \code{model} is a list of fitted detection functions, each model fitted to each replicate and number of adjustments is selected via AIC.
}
The last of these options can be very time consuming!
}

