% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootdht_Dhat_summarize.R
\name{bootdht_Dhat_summarize}
\alias{bootdht_Dhat_summarize}
\title{Simple summary of density results for bootstrap model}
\usage{
bootdht_Dhat_summarize(ests, fit)
}
\arguments{
\item{ests}{output from \code{\link{dht2}}.}

\item{fit}{fitted detection function object (unused).}
}
\value{
\code{data.frame} with two columns ("\code{Dhat}" and "\code{Label}"), giving the
estimate(s) of density of individuals per stratum from each bootstrap
replicate. This \code{data.frame} can be examined for example, with
\code{\link[stats:quantile]{quantile}} to compute confidence intervals.
}
\description{
When using \code{\link{bootdht}} one needs to use a summary function to
extract results from the resulting models per replicate. This function is
the simplest possible example of such a function, that just extracts the
estimated density (with stratum labels).
}
\details{
Further examples of such functions can be found at
\url{http://examples.distancesampling.org}.
}
\seealso{
\code{\link{bootdht}} which this function is to be used with and
\code{\link{bootdht_Nhat_summarize}} which does the same job
but returns abundance results.
}
