% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootdht_Nhat_summarize.R,
%   R/bootdht_Nhat_summary.R
\name{bootdht_Nhat_summarize}
\alias{bootdht_Nhat_summarize}
\title{Simple summary for bootstrap model}
\usage{
bootdht_Nhat_summarize(ests, fit)

bootdht_Nhat_summarize(ests, fit)
}
\arguments{
\item{ests}{output from \code{\link{dht2}}.}

\item{fit}{fitted detection function object (unused).}
}
\value{
\code{data.frame} with one column ("\code{Nhat}"), containing estimate(s) of abundundance of individuals.

\code{data.frame} with one column ("\code{Nhat}"), containing estimate(s) of abundundance of individuals.
}
\description{
When using \code{\link{bootdht}} one needs to use a summary function to extract results from the resulting models per replicate. This function is the simplest possible example of such a function, that just extracts the estimated abundance.

When using \code{\link{bootdht}} one needs to use a summary function to extract results from the resulting models per replicate. This function is the simplest possible example of such a function, that just extracts the estimated abundance.
}
\seealso{
\code{\link{bootdht}}, which this function is to be used with.

\code{\link{bootdht}}, which this function is to be used with.
}
