% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{capercaillie}
\alias{capercaillie}
\alias{capercaillie_units}
\title{Capercaillie in Monaughty Forest}
\format{
A \code{data.frame} with 112 observations on the following 9 variables.
\itemize{
\item \code{Sample.Label} name of single transect
\item \code{Effort} transect length (km)
\item \code{distance} perpendicular distance (m)
\item \code{object} object ID
\item \code{size} only individual birds detected
\item \code{detected} whether detected
\item \code{observer} single observer data
\item \code{Region.Label} stratum name
\item \code{Area} size of Monaughty Forest (ha)
}
}
\description{
Data from a line transect survey of capercaillie in Monaughty Forest, Moray,
Scotland.
}
\keyword{datasets}
