% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{Distance-package}
\alias{Distance}
\alias{Distance-package}
\title{Distance sampling}
\description{
\code{Distance} is a simple way to fit detection functions and estimate abundance using distance sampling methodology.
}
\details{
Underlying \code{Distance} is the package \code{mrds}, for more advanced analyses (such as those involving double observer surveys) one may find it necessary to use \code{mrds}.
}
\author{
David L. Miller <dave@ninepointeightone.net>
}
\references{
Laake, J.L. and D.L. Borchers. 2004. Methods for incomplete
  detection at distance zero. In: Advanced Distance Sampling, eds. S.T.
  Buckland, D.R.Anderson, K.P. Burnham, J.L. Laake, D.L. Borchers, and L.
  Thomas. Oxford University Press.

Marques, F.F.C. and S.T. Buckland. 2004. Covariate models for the detection
  function. In: Advanced Distance Sampling, eds. S.T. Buckland,
  D.R.Anderson, K.P. Burnham, J.L. Laake, D.L. Borchers, and L. Thomas.
  Oxford University Press.
}
\keyword{Models}
\keyword{Statistical}

