% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_E.R
\name{distance_E}
\alias{distance_E}
\title{Euclidean Distance of four values}
\usage{
distance_E(x1, x2, y1, y2)
}
\arguments{
\item{x1}{a numerical value which represents the x value of the first point.}

\item{x2}{a numerical value which represents the x value of the second point.}

\item{y1}{a numerical value which represents the y value of the first point.}

\item{y2}{a numerical value which represents the y value of the second point.}
}
\value{
A numerical value.
}
\description{
Euclidean Distance of four values
}
\examples{
x1 <- 2
x2 <- 8
y1 <- 5
y2 <- 1
distance_E(x1,x2,y1,y2)
}
