% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.gof.R
\name{ds.gof}
\alias{ds.gof}
\title{Goodness of fit tests for distance sampling models}
\usage{
ds.gof(model, breaks = NULL, nc = NULL, qq = TRUE, ks = FALSE, ...)
}
\arguments{
\item{model}{deprecated.}

\item{breaks}{deprecated.}

\item{nc}{deprecated.}

\item{qq}{deprecated.}

\item{ks}{deprecated.}

\item{\dots}{deprecated.}
}
\value{
Nothing, deprecated.
}
\description{
This function is deprecated, please see \code{\link{gof_ds}}.
}
\seealso{
\code{\link[mrds:qqplot.ddf]{qqplot.ddf}} \code{\link[mrds:ddf.gof]{ddf.gof}}
}
\author{
David L Miller
}
\keyword{utility}
