% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{DuikerCameraTraps}
\alias{DuikerCameraTraps}
\alias{DuikerCameraTraps_units}
\title{Duiker camera trap survey}
\format{
A \code{data.frame} with 6277 rows and 6 variables
\itemize{
\item \code{Region.Label}  strata names (single stratum)
\item \code{Area}  size of study area (40.37 km^2)
\item \code{multiplier}  spatial effort, as the proportion of a circle covered by
the angle of view of the camera (42 degrees for these cameras)
\item \code{Sample.Label}  camera station identifier (21 functioning cameras in
this data set)
\item \code{Effort}  temporal effort, i.e. the number of 2-second time-steps over
which the camera operated
\item \code{distance}  radial distance (m) to interval midpoint
}
}
\source{
Howe, E.J., Buckland, S.T., Després-Einspenner, M.-L. and Kühl, H.S.
(2017), Distance sampling with camera traps. Methods Ecol Evol, 8:
1558-1565. \doi{10.1111/2041-210X.12790}

Howe, Eric J. et al. (2018), Data from: Distance sampling with camera traps,
Dryad, Dataset, \doi{10.5061/dryad.b4c70}
}
\description{
Study took place in Tai National Park Cote d'Ivoire in 2014.  Filmed
Maxwell's duikers (Philantomba maxwellii) were assigned to distance
intervals; recorded distances are the midpoints of the intervals. This data
includes only observations recorded at times of peak activity.
}
\keyword{datasets}
