% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_ddf.R
\name{dummy_ddf}
\alias{dummy_ddf}
\title{Detection function objects when detection is certain}
\usage{
dummy_ddf(data, width, left = 0, transect = "line")
}
\arguments{
\item{data}{as specified for \code{ds} and \code{ddf}}

\item{width}{right truncation}

\item{left}{left truncation (default 0, no left truncation)}

\item{transect}{\code{"line"} or \code{"point"} transect}

\item{size}{group size for each observation (default all groups size 1)}
}
\description{
Create a detection function object for strip/plot surveys for use with
\code{dht2}.
}
\author{
David L Miller
}
