% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{ETP_Dolphin}
\alias{ETP_Dolphin}
\alias{ETP_Dolphin_units}
\title{Eastern Tropical Pacific spotted dolphin survey}
\format{
A \code{data.frame} with 1090 rows and 13 variables:
\itemize{
\item \code{Region.Label} stratum labels (only one)
\item \code{Area} size (nmi) of each stratum
\item \code{Sample.Label} transect labels
\item \code{Effort} transect length (nmi)
\item \code{object} object ID
\item \code{distance} perpendicular distance (nmi)
\item \code{LnCluster} natural log of cluster size
\item \code{Month} month of detection
\item \code{Beauf.class} Beaufort sea state
\item \code{Cue.type} initial cue triggering detection
\item \code{Search.method} observer method making the detection
\item \code{size} cluster size
\item \code{Study.Area} study area name
}
}
\source{
Inter-American Tropical Tuna Commission
}
\description{
Observers aboard tuna vessels detecting dolphin schools along with a number
of possibly useful covariates for modelling the detection function.
}
\details{
Several different search methods included in these data
\itemize{
\item \code{0} binoculars from crows nest
\item \code{2} binoculars from elsewhere on ship
\item \code{3} helicopter searching ahead of ship
\item \code{5} radar detects of seabirds above dolphin schools
}

Several cue types were also recorded by observers.
\itemize{
\item \code{1} seabirds above the school
\item \code{2} water splashes
\item \code{3} unspecified
\item \code{4} floating objects such as logs
}
}
\keyword{datasets}
