% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gof_ds.R
\name{gof_ds}
\alias{gof_ds}
\title{Goodness of fit testing and quantile-quantile plots}
\usage{
gof_ds(model, plot = TRUE, ...)
}
\arguments{
\item{model}{a fitted detection function.}

\item{plot}{if \code{TRUE} the Q-Q plot is plotted}
}
\description{
Computes chi-squared, Kolmogorov-Smirnov and Cramer-von Mises goodness of fit tests for the detection function. Optionally plot a quantile-quantile plot for fitted model as a graphical representation of goodness of fit.
}
\details{
See \code{\link{ddf.gof}} for details.
}

