% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.dsmodel.R
\name{logLik.dsmodel}
\alias{logLik.dsmodel}
\title{log-likelihood value for a fitted detection function}
\usage{
\method{logLik}{dsmodel}(object, ...)
}
\arguments{
\item{object}{a fitted detection function model object}

\item{\dots}{included for S3 completeness, but ignored}
}
\value{
a numeric value giving the log-likelihood with two attributes: \code{"df"} the "degrees of freedom" for the model (number of parameters) and \code{"nobs"} the number of observations used to fit the model
}
\description{
Extract the log-likelihood from a fitted detection function.
}
\examples{
\dontrun{
library(Distance)
data(minke)
model <- ds(minke, truncation=4)
# extract the log likelihood
logLik(model)
}
}
\author{
David L Miller
}
