% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{LTExercise}
\alias{LTExercise}
\alias{LTExercise_units}
\title{Simulated line transect survey data}
\format{A data frame with 106 rows and 7 variables
  \tabular{ll}{
    \code{Region.Label} \tab strata names (single stratum) \cr
    \code{Area} \tab size of study area (1 in this case, making abundance and density equal) \cr
    \code{Sample.Label} \tab transect ID \cr
    \code{Effort} \tab length of transects (km) \cr
    \code{object} \tab object ID  \cr
    \code{distance} \tab perpendicular distance (m) \cr
    \code{Study.Area} \tab name of study area \cr
  }}
\source{
Simulated data, from the distance sampling introductory course, Centre for
 Research into Ecological & Environmental Modelling, University of St Andrews.
}
\description{
Simulated line transect survey. Twelve transects, detection function is half-normal. 
True object density is 79.8 animals per km^2.
}
\note{
There is no unit object associated with this dataset
}
\keyword{datasets}
