% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_activity_fn.R
\name{make_activity_fn}
\alias{make_activity_fn}
\title{Multiplier bootstrap helper functions}
\usage{
make_activity_fn(...)
}
\value{
a function which generates a single bootstrap estimate of
availability
}
\description{
Use \link[activity:fitact]{activity::fitact} to fit an activity model.
}
\details{
Uses \link[activity:fitact]{activity::fitact} to generate single possible availability estimates
based on bootstraps. The function returns another function, which can be
passed to \code{bootdht}. It is recommended that you try out the function before
passing it to \code{bootdht}. See examples for a template for use.
}
\author{
David L Miller
}
