\name{plot.dsmodel}
\alias{plot.dsmodel}
\title{Plot a fitted detection function}
\usage{
  plot.dsmodel(x, ...)
}
\arguments{
  \item{x}{an object of class \code{dsmodel}.}
}
\value{
  \code{NULL}, just produces a plot.
}
\description{
  This is just a simple wrapper around
  \code{\link{plot.ds}}.
}
\author{
  David L. Miller
}

