% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dht_result.R
\name{print.dht_result}
\alias{print.dht_result}
\title{Print abundance estimates}
\usage{
\method{print}{dht_result}(x, report = "abundance", groups = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{dht_result}}

\item{report}{should \code{"abundance"}, \code{"density"} or \code{"both"} be reported?}

\item{groups}{should abundance/density of groups be produced? TRUE
by default.}

\item{\dots}{unused}
}
\description{
See \code{\link[=dht2]{dht2}} for information on printed column names.
}
