% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.dsmodel}
\alias{print.dsmodel}
\title{Simple pretty printer for distance sampling analyses}
\usage{
\method{print}{dsmodel}(x, ...)
}
\arguments{
\item{x}{a distance sampling analysis (result from calling \code{\link{ds}}).}

\item{...}{not passed through, just for S3 compatibility.}
}
\description{
Simply prints out a brief description of the model which was fitted. For more
detailed information use \code{\link{summary}}.
}
\author{
David L. Miller
}

