% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.dsmodel.R
\name{print.summary.dsmodel}
\alias{print.summary.dsmodel}
\title{Print summary of distance detection function model object}
\usage{
\method{print}{summary.dsmodel}(x, ...)
}
\arguments{
\item{x}{a summary of distance sampling analysis}

\item{\dots}{unspecified and unused arguments for S3 consistency}
}
\value{
Nothing, just prints the summary.
}
\description{
Provides a brief summary of a distance sampling analysis. Including: detection 
function parameters, model selection criterion, and optionally abundance in the
covered (sampled) region and its standard error.
}
\seealso{
\code{\link{summary.ds}}
}
\author{
David L. Miller and Jeff Laake
}
\keyword{utility}
