% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{PTExercise}
\alias{PTExercise}
\alias{PTExercise_units}
\title{Simulated point transect survey data}
\format{
A \code{data.frame} with 144 rows and 7 variables
\itemize{
\item \code{Region.Label} strata names (single stratum)
\item \code{Area} size of study area (0 in this case)
\item \code{Sample.Label} transect ID
\item \code{Effort} number of visits to point
\item \code{object} object ID
\item \code{distance} radial distance (m)
\item \code{Study.Area} name of study area
}
}
\source{
Simulated data, from the distance sampling introductory course,
Centre for Research into Ecological & Environmental Modelling, University of
St Andrews.
}
\description{
Simulated point transect survey. Thirty point transects, detection function
is half-normal.  True object density is 79.6 animals per hectare.
}
\keyword{datasets}
