% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAIC.dsmodel.R
\name{QAIC.dsmodel}
\alias{QAIC.dsmodel}
\title{QIAC Information Criterion for detection functions when data is overdispersed}
\usage{
QAIC.dsmodel(object, ...)
}
\arguments{
\item{object}{a fitted detection function object}

\item{\dots}{optionally more fitted model objects.}
}
\description{
Overdispersion causes AIC to select overly-complex models, so analysts
should specify the number/order of adjustment terms manually when fitting
distance sampling models to data from camera traps, rather than allowing
automated selection using AIC. Howe et al (2019) describes two methods for
performing model selection of distance sampling models in the face of
overdispersion.
}
\details{
The first method of Howe et al (2019) employs a two-step process. First, an
overdisersion factor (\code{chat}) is computed for each key function family from
the most complex model in each family (derived from the
chi-squared goodness of fit test statistic divided by its degrees of
freedom).
}
\references{
Howe, E. J., Buckland, S. T., Després-Einspenner, M.-L., & Kühl, H. S. (2019). Model selection with overdispersed distance sampling data. Methods in Ecology and Evolution, 10(1), 38–47. \doi{10.1111/2041-210X.13082}
}
\author{
David L Miller, based on code from Eric Rexstad
}
