% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAIC.R
\name{QAIC}
\alias{QAIC}
\alias{chi2_select}
\title{Tools for model selection when distance sampling data are overdispersed}
\usage{
QAIC(object, ..., chat = NULL, k = 2)

chi2_select(object, ...)
}
\arguments{
\item{object}{a fitted detection function object}

\item{\dots}{additional fitted model objects.}

\item{chat}{a value of \eqn{\hat{c}}{chat} to be used in QIAC calculation}

\item{k}{penalty per parameter to be used; default 2}
}
\value{
a \code{data.frame} with one row per model supplied, in the same order as
given
}
\description{
Overdispersion causes AIC to select overly-complex models, so analysts
should specify the number/order of adjustment terms manually when fitting
distance sampling models to data from camera traps, rather than allowing
automated selection using AIC. Howe et al (2019) describes two methods for
performing model selection of distance sampling models in the face of
overdispersion. In both cases comparisons should be made between models
which use the same key function in their detection function specification.
}
\details{
The first method of Howe et al (2019) employs a two-step process:
\enumerate{
\item An overdisersion factor (\eqn{\hat{c}_1}{chat1}) is computed for each
key function family from the most complex model in each family (derived
from the chi-squared goodness of fit test statistic divided by its degrees
of freedom). When multiple detection functions are supplied to \code{QAIC},
\code{chat} will be calculated automatically. Alternatively \code{chat} can be
supplied as an argument to \code{QAIC}.
\item To compare between key function classes, the chi-squared goodness of fit
test statistic divided by its degrees is used. The function \code{chi2_select}
performs this calculation.
An example of this process is given in the examples below.
}

The second method (\eqn{\hat{c}_2}{chat2}) in the notation of Howe et al)
uses the number of distance observations recorded per independent encounter
between an animal and an observer. This quantity can be calculated from the
raw data. In camera trap surveys of solitary animals,
\eqn{\hat{c}_2}{chat2} would be the mean number of distance observations
recorded during a single pass by an animal in front of a trap. In surveys of
social animals employing human observers, \eqn{\hat{c}_2}{chat2} would be
the mean number of detected animals per detected group, and in camera trap
surveys of social animals \eqn{\hat{c}_2}{chat2} the mean number of
distance observations recorded during an encounter between a group of
animals and a CT. In this case one can calculate \eqn{\hat{c}_2}{chat2} and
use the \code{chat} argument to specify it for one or more models.
}
\examples{
\dontrun{
library(Distance)
data("wren_cuecount")

# fit hazard-rate key models
w3.hr0 <- ds(wren_cuecount, transect="point", key="hr", adjustment=NULL,
             truncation=92.5)
w3.hr1 <- ds(wren_cuecount, transect="point", key="hr", adjustment="cos",
             order=2, truncation=92.5)
w3.hr2 <- ds(wren_cuecount, transect="point", key="hr", adjustment="cos",
             order=c(2, 4), truncation=92.5)

# fit unform key models
w3.u1 <- ds(wren_cuecount, transect="point", key="unif", adjustment="cos",
            order=1, truncation=92.5)
w3.u2 <- ds(wren_cuecount, transect="point", key="unif", adjustment="cos",
            order=c(1,2), monotonicity="none",  truncation=92.5)
w3.u3 <- ds(wren_cuecount, transect="point", key="unif", adjustment="cos",
            order=c(1,2,3), monotonicity="none", truncation=92.5)

# fit half-normal key functions
w3.hn0 <- ds(wren_cuecount, transect="point", key="hn", adjustment=NULL,
             truncation=92.5)
w3.hn1 <- ds(wren_cuecount, transect="point", key="hn", adjustment="herm",
             order=2, truncation=92.5)

# stage 1: calculate QAIC per model set
QAIC(w3.hr0, w3.hr1, w3.hr2)  # no adjustments smallest
QAIC(w3.u1, w3.u2, w3.u3)     # 2 adjustment terms (by 0.07)
QAIC(w3.hn0, w3.hn1)  # no adjustments smallest

# stage 2: select using chi^2/degrees of freedom between sets
chi2_select(w3.hr0, w3.u2, w3.hn0)

# exmaple using a pre-calculated chat
chat <- attr(QAIC(w3.hr0, w3.hr1, w3.hr2), "chat")
QAIC(w3.hr0, chat=chat)
}
}
\references{
Howe, E. J., Buckland, S. T., Després-Einspenner, M.-L., & Kühl, H. S. (2019). Model selection with overdispersed distance sampling data. Methods in Ecology and Evolution, 10(1), 38–47. \doi{10.1111/2041-210X.13082}
}
\author{
David L Miller, based on code from Eric Rexstad and explanation from
Eric Howe.
}
