% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_factorize.R
\name{safe_factorize}
\alias{safe_factorize}
\title{Safely make the right things factors in data.frames}
\usage{
safe_factorize(formula, data)
}
\arguments{
\item{formula}{an R formula}

\item{data}{a data.frame}
}
\value{
data, with (maybe) some extra columns
}
\description{
If you have a formula with as.factor(var) in it, this function will create a new column with the factor version of var with a column called as.factor(var). Uses model.frame
}
\author{
David L Miller
}
