% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{Savannah_sparrow_1980}
\alias{Savannah_sparrow_1980}
\alias{Savannah_sparrow_1981}
\alias{Savannah_sparrow_1980_units}
\alias{Savannah_sparrow_1981_units}
\title{Savanna sparrow point transects}
\format{\code{data.frame} with 468 observations of 7 variables:
\tabular{ll}{
\code{Region.Label} \tab stratum label (pasture ID) \cr
\code{Area} \tab stratum area (set to 1 so density is reported) \cr
\code{Sample.Label} \tab transect identifier \cr
\code{Effort} \tab number of visits \cr
\code{object} \tab object ID \cr
\code{distance} \tab radial distance (m) \cr
\code{Study.Area} \tab name of study area
}}
\description{
Point transect data collected in Colorado 1980/81 to examine effect of
agricultural practices upon avian community.
}
\details{
Design consisted of point transects placed in multiple pastures 
(3 in 1980, 4 in 1981).  While many species were observed, only data for 
Savannah sparrows (Passerculus sandwichensis) are included here.
}
\note{
Data structure for 1981 data set is identical, but there are 
448 observations across 4 pastures.
}
\references{
Knopf, F.L., J.A. Sedgwick, and R.W. Cannon. (1988) Guild 
structure of a riparian avifauna relative to seasonal cattle grazing. 
The Journal of Wildlife Management 52 (2): 280–290. https://doi.org/10.2307/3801235.
}
\keyword{datasets}
