% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{sikadeer}
\alias{sikadeer}
\alias{sikadeer_units}
\title{Sika deer pellet data from southern Scotland}
\format{
A data frame with 1923 rows and 11 variables.
  \tabular{ll}{  
    \code{Region.Label} \tab stratum labels \cr
    \code{Area} \tab size (ha) of each stratum \cr
    \code{Sample.Label} \tab transect labels \cr
    \code{Defecation.rate} \tab rate of dung production per individual per day \cr
    \code{Defecation.rate.SE} \tab variability in defecation rate \cr
    \code{Decay.rate} \tab time (days) for dung to become undetectable \cr
    \code{Decay.rate.SE} \tab variability in decay rate \cr
    \code{Effort} \tab transect length (km) \cr
    \code{object} \tab object ID \cr
    \code{distance} \tab perpendicular distance (cm) \cr
    \code{Study.Area} \tab study area name \cr
  }
}
\description{
Because sika deer spend most of their time in woodland areas, abundance
estimates are based on pellet group counts.  Line transect methods were applied
to estimate deer pellet group density by geographic block.
}
\details{
Data presented here are from the Peebleshire portion of the study described by 
Marques et al. (2001).
}
\references{
Marques, F.F.C., S.T. Buckland, D. Goffin, C.E. Dixon, D.L. Borchers, B.A. Mayle, and A.J. Peace. (2001).
 Estimating deer abundance from line transect surveys of dung: 
 sika deer in southern Scotland. Journal of Applied Ecology 38 (2): 349–363.
  https://doi.org/10.1046/j.1365-2664.2001.00584.x
}
\keyword{datasets}
