% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarize_ds_models.R
\name{summarize_ds_models}
\alias{summarize_ds_models}
\title{Make a table of summary statistics for detection function models}
\usage{
summarize_ds_models(..., sort = "AIC")
}
\arguments{
\item{...}{models to be summarised}

\item{sort}{column to sort by (default \code{"AIC"})}
}
\description{
Provide a summary table of useful information about fitted detection functions. This can be useful when paired with \code{knitr}s \code{\link{kable}} function.
}
\author{
David L Miller
}

