% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dht_bootstrap.R
\name{summary.dht_bootstrap}
\alias{summary.dht_bootstrap}
\title{Summarize bootstrap abundance uncertainty estimate output}
\usage{
\method{summary}{dht_bootstrap}(object, alpha = 0.05, ...)
}
\arguments{
\item{object}{output from \code{bootdht}}

\item{alpha}{value to use in confidence interval calculation (to obtain \code{alpha}/2 and 1-\code{alpha}/2 intervals}

\item{...}{for S3 compatibility, unused.}
}
\value{
a \code{data.frame} of summary statistics
}
\description{
A simple function to calculate summaries of bootstrap output generated by \code{\link{bootdht}}.
}
\details{
Summaries are only made for numeric outputs. Both median and mean are reported to allow assessment of bias. The coefficient of variation reported (in column \code{cv}) is based on the median calculated from the bootstraps.
}
