% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{Systematic_variance_1}
\alias{Systematic_variance_1}
\alias{Systematic_variance_2}
\alias{Systematic_variance_1_units}
\alias{Systematic_variance_2_units}
\title{Simulation of encounter rate variance}
\format{
\code{data.frame} with 253 observations (\code{systematic_var_1}) or 256
observations (\code{systematic_var_2}) of 7 variables:
\code{Region.Label} stratum label (default)
\code{Area} stratum area (0.5 km^2)
\code{Sample.Label} transect identifier
\code{Effort} transect length (km)
\code{object} object ID
\code{distance} perpendicular distance (m)
\code{Study.Area} name of study area
}
\description{
\code{systematic_var_1} consists of simulated line transect data with large
differences in transect length. In \code{systematic_var_2} that transect length
gradient is coupled with a strong animal gradient; exaggerating encounter
rate variance between transects.
}
\details{
True population size is 1000 objects in the study area of size 0.5 km^2;
such that true density is 2000 objects per km.
}
\references{
Fewster, R.M., S.T. Buckland, K.P. Burnham, D.L. Borchers, P.E.
Jupp, J.L. Laake and L. Thomas. (2009) Estimating the encounter rate
variance in distance sampling. Biometrics 65 (1): 225–236.
\doi{10.1111/j.1541-0420.2008.01018.x}
}
\keyword{datasets}
