% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unflatten.R
\name{unflatten}
\alias{unflatten}
\title{Unflatten flatfile data.frames}
\usage{
unflatten(data)
}
\arguments{
\item{data}{data in flatfile format (a \code{data.frame})}
}
\value{
\code{list} of four \code{data.frame}s: distance data, observation table,
sample table, region table.
}
\description{
Sometimes data is provided in the \code{\link{flatfile}} format, but we
really want it in \code{mrds} format (that is, as distance data, observation
table, sample table and region table format). This function undoes the
flattening, assuming that the data have the correct columns.
}
\author{
David L Miller
}
