% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{unimak}
\alias{unimak}
\alias{unimak_units}
\title{Simulated line transect survey data with covariates}
\format{
A \code{data.frame} with 60 rows and 9 variables
\itemize{
\item \code{Region.Label} strata names (single stratum)
\item \code{Area} size of study area (mi^2)
\item \code{Sample.Label} transect ID
\item \code{Effort} transect length (mi)
\item \code{object} object ID
\item \code{distance} perpendicular distance (km)
\item \code{MSTDO} time since medication taken by observer (min)
\item \code{Hour} time of day of sighting (hour)
\item \code{Study.Area} name of study area
}
}
\source{
Simulated data, from the distance sampling introductory course,
Centre for Research into Ecological & Environmental Modelling, University of
St Andrews.
}
\description{
Simulated line transect survey. Only eight line transects, detection
function is half-normal.
}
\note{
\code{Hour} is covariate that has no effect on detection function,
while \code{MSTDO} does affect the detection function.  Examine the ability
of model selection to choose the correct model.
}
\keyword{datasets}
