% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_table.R
\name{units_table}
\alias{units_table}
\title{Generate table of unit conversions}
\usage{
units_table()
}
\description{
Returns a table of conversions between the units used in Distance for Windows. This is extracted from the \code{DistIni.mdb} default database.
}
\author{
David L Miller
}
