% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.dsmodel.R
\name{AIC.dsmodel}
\alias{AIC.dsmodel}
\title{Akaike's An Information Criterion for detection functions}
\usage{
\method{AIC}{dsmodel}(object, ..., k = 2)
}
\arguments{
\item{object}{a fitted detection function object}

\item{\dots}{required for S3 but ignored}

\item{k}{penalty per parameter to be used; the default \code{k = 2} is the "classical" AIC}
}
\description{
Extract the AIC from a fitted detection function.
}
\author{
David L Miller
}
