% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{amakihi}
\alias{amakihi}
\alias{amakihi_units}
\title{Hawaiian amakihi point transect data}
\format{
A \code{data.frame} with 1487 rows and 12 variables
\itemize{
\item \code{Region.Label} strata names (seven strata)
\item \code{Area} size of study area (set to 0)
\item \code{Sample.Label} transect ID
\item \code{Effort} number of visits to point
\item \code{object} object ID
\item \code{distance} radial distance (m)
\item \code{Month} month survey conducted (not used)
\item \code{OBs} observer ID (note capitalisation of variable name)
\item \code{Sp} species code (COAM) for all detections
\item \code{MAS} Time after sunrise (min)
\item \code{HAS} Time after sunrise (hours)
\item \code{Study.Area} name of study area
}
}
\description{
Also known as the Common 'Amakihi, a type of Hawaiian honeycreeper
}
\note{
Example for investigating covariates in the detection function.  Note
high colinearity between two measures of time since sunrise.  Convergence
problems can result from models with several factor covariates.
}
\references{
Marques, T.A., L. Thomas, S.G. Fancy and S.T. Buckland. (2007)
Improving estimates of bird density using multiple-covariate distance
sampling.  The Auk 124 (4): 1229–1243.
\doi{10.1642/0004-8038(2007)124[1229:IEOBDU]2.0.CO;2}
}
\keyword{datasets}
