% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amakihi.R
\docType{data}
\name{amakihi}
\alias{amakihi}
\title{Amakihi (Hemignathus virens) point transect data}
\format{1485 observations of 7 variables: \code{survey}, \code{object}, \code{distance}, \code{obs}, \code{mas}, \code{has}, \code{detected}.}
\description{
1485 observations of Hawaiian amakihi. Data collected on Hawaii from point transects, collected as part of a larger study to assess a Palila (Loxioides bailleuli) translocation experiment on the island of Hawaii (Fancy et al. 1997) between July 1992 and April 1995 (7 survey periods). Full analyses of the amakihi data is provided in Marques et al (2007).
}
\details{
Data include distances, as well as survey period (\code{survey}, factor), observer code (\code{obs}, factor), hours after sunrise (\code{has}, factor) and minutes after sunrise (\code{mas}).

We thank Steve Fancy for making this data publicly available.
}
\references{
Fancy, SG, TJ Snetsinger, & JD Jacobi (1997). Translocation of the Palila, an Endangered Hawaiian Honeycreeper. Pacific Conservation, 3(1).

Marques, TA, L Thomas, S Fancy & ST Buckland (2007). Improving estimates of bird density using multiple-covariate distance sampling. The Auk, 124(4), 1229-1243.
}
\keyword{datasets}
