% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{capercaillie}
\alias{capercaillie}
\alias{capercaillie_units}
\title{Capercaillie in Monaughty Forest}
\format{A data frame with 112 observations on the following 9 variables.
\tabular{ll}{
  \code{Sample.Label} \tab name of single transect \cr
  \code{Effort} \tab transect length (km) \cr
  \code{distance} \tab perpendicular distance (m) \cr
  \code{object} \tab object ID \cr
  \code{size} \tab only individual birds detected \cr
  \code{detected} \tab whether detected \cr
  \code{observer} \tab single observer data \cr
  \code{Region.Label} \tab stratum name \cr
  \code{Area} \tab size of Monaughty Forest (ha) \cr
}}
\description{
Data from a line transect survey of capercaillie in Monaughty Forest, Moray, 
Scotland.
}
\keyword{datasets}
