% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkdata.R
\name{checkdata}
\alias{checkdata}
\title{Check that the data supplied to \code{ds} is correct}
\usage{
checkdata(data, region.table = NULL, sample.table = NULL,
  obs.table = NULL, formula = ~1)
}
\arguments{
\item{data}{as in \code{ds}}

\item{region.table}{as in \code{ds}}

\item{sample.table}{as in \code{ds}}

\item{obs.table}{as in \code{ds}}

\item{formula}{formula for the covariates}
}
\value{
Throws an error if something goes wrong, otherwise returns a
 \code{data.frame}.
}
\description{
This is an internal function that checks the \code{data.frame}s supplied
to \code{ds} are "correct".
}
\author{
David L. Miller
}
