% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ds.gof.R
\name{ds.gof}
\alias{ds.gof}
\title{Goodness of fit tests for distance sampling models}
\usage{
ds.gof(model, breaks = NULL, nc = NULL, qq = TRUE, ...)
}
\arguments{
\item{model}{fitted model object}

\item{breaks}{Cutpoints to use for binning data}

\item{nc}{Number of distance classes}

\item{qq}{Flag to indicate whether quantile-quantile plot is desired}

\item{\dots}{Graphics parameters to pass into qqplot function}
}
\value{
List of test results and a plot.
}
\description{
Chi-square, Kolmogorov-Smirnov and Cramer-von Mises goodness of fit tests for detection function models.
}
\author{
David L Miller
}
\seealso{
qqplot.ddf ddf.gof
}
\keyword{utility}

