% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.gof.R
\name{ds.gof}
\alias{ds.gof}
\title{Goodness of fit tests for distance sampling models}
\usage{
ds.gof(model, breaks = NULL, nc = NULL, qq = TRUE, ks = FALSE, ...)
}
\arguments{
\item{model}{fitted model object}

\item{breaks}{Cutpoints to use for binning data}

\item{nc}{Number of distance classes}

\item{qq}{Flag to indicate whether quantile-quantile plot is desired}

\item{ks}{perform the Kolmogorov-Smirnov test (this involves many bootstraps so can take a while)}

\item{\dots}{Graphics parameters to pass into qqplot function}
}
\value{
List of test results and a plot.
}
\description{
Chi-square, Kolmogorov-Smirnov (if \code{ks=TRUE}) and Cramer-von Mises goodness of fit tests for detection function models.
}
\examples{
\dontrun{
# fit and test a simple model for the golf tee data
library(Distance)
data(book.tee.data)
tee.data<-book.tee.data$book.tee.dataframe[book.tee.data$book.tee.dataframe$observer==1,]
ds.model <- ds(tee.data,4)
ds.gof(ds.model)
}
}
\seealso{
qqplot.ddf ddf.gof
}
\author{
David L Miller
}
\keyword{utility}
