% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{ds}
\alias{ds}
\title{Fit detection functions and calculate abundance from line or point transect data}
\usage{
ds(data, truncation = ifelse(is.null(cutpoints), ifelse(is.null(data$distend),
  max(data$distance), max(data$distend)), max(cutpoints)),
  transect = c("line", "point"), formula = ~1, key = c("hn", "hr",
  "unif"), adjustment = c("cos", "herm", "poly"), order = NULL,
  scale = c("width", "scale"), cutpoints = NULL, dht.group = FALSE,
  monotonicity = ifelse(formula == ~1, "strict", "none"),
  region.table = NULL, sample.table = NULL, obs.table = NULL,
  convert.units = 1, method = "nlminb", quiet = FALSE, debug.level = 0,
  initial.values = NULL)
}
\arguments{
\item{data}{a \code{data.frame} containing at least a column called
\code{distance} or a numeric vector containing the distances. 
NOTE! If there is a column called \code{size} in
the data then it will be interpreted as group/cluster size, see the
section "Clusters/groups", below. One can supply data as a "flat file"
and not supply \code{region.table}, \code{sample.table} and
\code{obs.table}, see "Data format", below and \code{\link{flatfile}}.}

\item{truncation}{either truncation distance (numeric, e.g. 5) or percentage (as a string, e.g. "15\%"). Can be supplied as a \code{list} with elements \code{left} and \code{right} if left truncation is required (e.g. \code{list(left=1,right=20)} or \code{list(left="1\%",right="15\%")} or even \code{list(left="1",right="15\%")}).
By default for exact distances the maximum observed distance is used as the right truncation. When the data is binned, the right truncation is the largest bin end point. Default left truncation is set to zero.}

\item{transect}{indicates transect type "line" (default) or "point".}

\item{formula}{formula for the scale parameter. For a CDS analysis leave this as its default \code{~1}.}

\item{key}{key function to use; "hn" gives half-normal (default), "hr" gives hazard-rate and "unif" gives uniform. Note that if uniform key is used, covariates cannot be included in the model.}

\item{adjustment}{adjustment terms to use; \code{"cos"} gives cosine (default), \code{"herm"} gives Hermite polynomial and \code{"poly"} gives simple polynomial. \code{"cos"} is recommended. A value of \code{NULL} indicates that no adjustments are to be fitted.}

\item{order}{orders of the adjustment terms to fit (as a vector/scalar), the default value (\code{NULL}) will select via AIC up to order 5. If a single number is given, that number is expanded to be \code{seq(term_min, order, by=1)} where \code{term.min} is the appropriate minimum order for this type of adjustment. For cosine adjustments, valid orders are integers greater than 2 (except when a uniform key is used, when the minimum order is 1). For Hermite polynomials, even integers equal or greater than 2 are allowed and for simple polynomials even integers equal or greater than 2 are allowed (though note these will be multiplied by 2, see Buckland et al, 2001 for details on their specification). By default, AIC selection will try up to 5 adjustments, beyond that you must specify these manually, e.g. \code{order=2:6} and perform your own AIC selection.}

\item{scale}{the scale by which the distances in the adjustment terms are divided. Defaults to \code{"width"}, scaling by the truncation distance. If the key is uniform only \code{"width"} will be used. The other option is \code{"scale"}: the scale parameter of the detection}

\item{cutpoints}{if the data are binned, this vector gives the cutpoints of the bins. Ensure that the first element is 0 (or the left truncation distance) and the last is the distance to the end of the furthest bin. (Default \code{NULL}, no binning.) Note that if \code{data} has columns \code{distbegin} and \code{distend} then these will be used as bins if \code{cutpoints} is not specified. If both are specified, \code{cutpoints} has precedence.}

\item{dht.group}{should density abundance estimates consider all groups to be size 1 (abundance of groups) \code{dht.group=TRUE} or should the abundance of individuals (group size is taken into account), \code{dht.group=FALSE}. Default is \code{FALSE} (abundance of individuals is calculated).}

\item{monotonicity}{should the detection function be constrained for monotonicity weakly (\code{"weak"}), strictly (\code{"strict"}) or not at all (\code{"none"} or \code{FALSE}). See Montonicity, below. (Default \code{"strict"}). By default it is on for models without covariates in the detection function, off when covariates are present.}

\item{region.table}{\code{data.frame} with two columns:
\tabular{ll}{ \code{Region.Label} \tab label for the region\cr
             \code{Area} \tab area of the region\cr}
\code{region.table} has one row for each stratum. If there is no stratification then \code{region.table} has one entry with \code{Area} corresponding to the total survey area.}

\item{sample.table}{\code{data.frame} mapping the regions to the samples (i.e. transects). There are three columns:
\tabular{ll}{\code{Sample.Label} \tab label for the sample\cr
             \code{Region.Label} \tab label for the region that the
                  sample belongs to.\cr
             \code{Effort} \tab the effort expended in that sample
                  (e.g. transect length).\cr}}

\item{obs.table}{\code{data.frame} mapping the individual observations (objects) to regions and samples. There should be three columns:
\tabular{ll}{\code{object} \tab unique numeric identifier for the observation\cr
             \code{Region.Label} \tab label for the region that the
                  sample belongs to.\cr
             \code{Sample.Label} \tab label for the sample\cr}}

\item{convert.units}{conversion between units for abundance estimation, see "Units", below. (Defaults to 1, implying all of the units are "correct" already.)}

\item{method}{optimization method to use (any method usable by \code{\link{optim}} or \pkg{optimx}). Defaults to \code{"nlminb"}.}

\item{quiet}{surpress non-essential messages (useful for bootstraps etc). Default value \code{FALSE}.}

\item{debug.level}{print debugging output. \code{0}=none, \code{1-3} increasing levels of debugging output.}

\item{initial.values}{a \code{list} of named starting values, see \code{\link{mrds-opt}}. Only allowed when AIC term selection is not used.}
}
\value{
a list with elements:
       \tabular{ll}{\code{ddf} \tab a detection function model object.\cr
                    \code{dht} \tab abundance/density information (if survey
                     region data was supplied, else \code{NULL}).}
}
\description{
This function fits detection functions to line or point transect data and then (provided that survey information is supplied) calculates abundance and density estimates. The examples below illustrate some basic types of analysis using \code{ds()}.
}
\section{Details}{

If abundance estimates are required then the \code{data.frame}s \code{region.table} and \code{sample.table} must be supplied. If \code{data} does not contain the columns \code{Region.Label} and \code{Sample.Label} then the \code{data.frame} \code{obs.table} must also be supplied. Note that stratification only applies to abundance estimates and not at the detection function level.
}

\section{Clusters/groups}{

 Note that if the data contains a column named \code{size} and \code{region.table}, \code{sample.table} and \code{obs.table} are supplied, cluster size will be estimated and density/abundance will be based on a clustered analsis of the data. Setting this column to be \code{NULL} will perform a non-clustred analysis (for example if "\code{size}" means something else in your dataset).
}

\section{Truncation}{

The right truncation point is by default set to be largest observed distance or bin end point. This is a default will not be appropriate for all data and can often be the cause of model convergence failures. It is recommended that one plots a histogram of the observed distances prior to model fitting so as to get a feel for an appropriate truncation distance. (Similar arguments go for left truncation, if appropriate). Buckland et al (2001) provide guidelines on truncation.

When specified as a percentage, the largest \code{right} and smallest \code{left} percent distances are discarded. Percentages cannot be supplied when using binned data.

For left truncation, there are two options: (1) fit a detection function to the truncated data as is (this is what happens when you set \code{left}). This does not assume that g(x)=1 at the truncation point. (2) manually remove data with distances less than the left truncation distance -- effectively move the centreline out to be the truncation distance (this needs to be done before calling \code{ds}). This then assumes that detection is certain at the left truncation distance. The former strategy has a weaker assumption, but will give higher variance as the detection function close to the line has no data to tell it where to fit -- it will be relying on the data from after the left truncation point and the assumed shape of the detection function. The latter is most appropriate in the case of aerial surveys, where some area under the plane is not visible to the observers, but their probability of detection is certain at the smallest distance.

 @section Binning: Note that binning is performed such that bin 1 is all distances greater or equal to cutpoint 1 (>=0 or left truncation distance) and less than cutpoint 2. Bin 2 is then distances greater or equal to cutpoint 2 and less than cutpoint 3 and so on.
}

\section{Monotonicity}{
 When adjustment terms are used, it is possible for the detection function to not always decrease with increasing distance. This is unrealistic and can lead to bias. To avoid this, the detection function can be constrained for monotonicity (and is by default for detection functions without covariates).

 Monotonicity constraints are supported in a similar way to that described in Buckland et al (2001). 20 equally spaced points over the range of the detection function (left to right truncation) are evaluated at each round of the optimisation and the function is constrained to be either always less than it's value at zero (\code{"weak"}) or such that each value is less than or equal to the previous point (monotonically decreasing; \code{"strict"}). See also \code{\link{check.mono}} in \code{mrds}.

Even with no monotonicity constraints, checks are still made that the detection function is monotonic, see \code{\link{check.mono}}.
}

\section{Units}{

 In extrapolating to the entire survey region it is important that
 the unit measurements be consistent or converted for consistency.
 A conversion factor can be specified with the \code{convert.units}
 variable.  The values of \code{Area} in \code{region.table}, must be made
 consistent with the units for \code{Effort} in \code{sample.table} and the
 units of \code{distance} in the \code{data.frame} that was analyzed.  It is
 easiest if the units of \code{Area} are the square of the units of
 \code{Effort} and then it is only necessary to convert the units of
 \code{distance} to the units of \code{Effort}. For example, if \code{Effort}
  was entered in kilometers and \code{Area} in square kilometers and
 \code{distance} in meters then using \code{convert.units=0.001} would
 convert meters to kilometers, density would be expressed in square
 kilometers which would then be consistent with units for \code{Area}.
 However, they can all be in different units as long as the appropriate
 composite value for \code{convert.units} is chosen.  Abundance for a survey
 region can be expressed as: \code{A*N/a} where \code{A} is \code{Area} for
 the survey region, \code{N} is the abundance in the covered (sampled)
 region, and \code{a} is the area of the sampled region and is in units of
 \code{Effort * distance}.  The sampled region \code{a} is multiplied by
  \code{convert.units}, so it should be chosen such that the result is in
 the same units as \code{Area}.  For example, if \code{Effort} was entered
 in kilometers, \code{Area} in hectares (100m x 100m) and \code{distance}
 in meters, then using \code{convert.units=10} will convert \code{a} to
 units of hectares (100 to convert meters to 100 meters for distance and
 .1 to convert km to 100m units).

 @section Data format: One can supply \code{data} only to simply fit a detection function. However, if abundance/density estimates are necessary further information is required. Either the \code{region.table}, \code{sample.table} and \code{obs.table} \code{data.frame}s can be supplied or all data can be supplied as a "flat file" in the \code{data} argument. In this format each row in data has additional information that would ordinarily be in the other tables. This usually means that there are additional columns named: \code{Sample.Label}, \code{Region.Label}, \code{Effort} and \code{Area} for each observation. See \code{\link{flatfile}} for an example.
}

\examples{

# An example from mrds, the golf tee data.
library(Distance)
data(book.tee.data)
tee.data<-book.tee.data$book.tee.dataframe[book.tee.data$book.tee.dataframe$observer==1,]
ds.model <- ds(tee.data,4)
summary(ds.model)
plot(ds.model)

\dontrun{
# same model, but calculating abundance
# need to supply the region, sample and observation tables
region <- book.tee.data$book.tee.region
samples <- book.tee.data$book.tee.samples
obs <- book.tee.data$book.tee.obs

ds.dht.model <- ds(tee.data,4,region.table=region,
             sample.table=samples,obs.table=obs)
summary(ds.dht.model)

# specify order 2 cosine adjustments
ds.model.cos2 <- ds(tee.data,4,adjustment="cos",order=2)
summary(ds.model.cos2)

# specify order 2 and 3 cosine adjustments, turning monotonicity
# constraints off
ds.model.cos23 <- ds(tee.data,4,adjustment="cos",order=c(2,3),
                   monotonicity=FALSE)
# check for non-monotonicity -- actually no problems
check.mono(ds.model.cos23$ddf,plot=TRUE,n.pts=100)

# include both a covariate and adjustment terms in the model
ds.model.cos2.sex <- ds(tee.data,4,adjustment="cos",order=2,
                        monotonicity=FALSE, formula=~as.factor(sex))
# check for non-monotonicity -- actually no problems
check.mono(ds.model.cos2.sex$ddf,plot=TRUE,n.pts=100)

# truncate the largest 10\% of the data and fit only a hazard-rate
# detection function
ds.model.hr.trunc <- ds(tee.data,truncation="10\%",key="hr",adjustment=NULL)
summary(ds.model.hr.trunc)

# compare AICs between these models:
AIC(ds.model)
AIC(ds.model.cos2)
AIC(ds.model.cos23)
}

}
\references{
Buckland, S.T., Anderson, D.R., Burnham, K.P., Laake, J.L., Borchers, D.L., and Thomas, L. (2001). Distance Sampling. Oxford University Press. Oxford, UK.

Buckland, S.T., Anderson, D.R., Burnham, K.P., Laake, J.L., Borchers, D.L., and Thomas, L. (2004). Advanced Distance Sampling. Oxford University Press. Oxford, UK.
}
\seealso{
\code{\link{flatfile}}
}
\author{
David L. Miller
}
