% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{ducknest}
\alias{ducknest}
\alias{ducknest_units}
\alias{ducknests_units}
\title{Ducknest line transect survey data}
\format{A data frame with 534 rows and 7 variables
  \tabular{ll}{
    \code{Region.Label} \tab strata names (single stratum in this instance) \cr
    \code{Area} \tab size of refuge (0 in this case, actual size 60km^2) \cr
    \code{Sample.Label} \tab transect ID \cr
    \code{Effort} \tab length of transects (km) \cr
    \code{object} \tab nest ID  \cr
    \code{distance} \tab perpendicular distance (m) \cr
    \code{Study.Area} \tab name of wildlife refuge \cr
  }}
\source{
Simulated data, from the distance sampling introductory course, Centre for Research into Ecological & Environmental Modelling, University of St Andrews.
}
\description{
Simulated line transect survey of duck nests, designed to reproduce the data of Figure 2 in Anderson and Pospahala (1970).
}
\details{
The Monte Vista National Wildlife Refuge is in southern Colorado in the USA at an altitude
of roughly 2400m.
}
\references{
Anderson, D. R., and R. S. Pospahala. 1970. Correction of bias in belt transect studies of immotile objects. The Journal of Wildlife Management 34 (1): 141–146. https://doi.org/10.2307/3799501
}
\keyword{datasets}
