% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_ddf.R
\name{dummy_ddf}
\alias{dummy_ddf}
\title{Detection function objects when detection is certain}
\usage{
dummy_ddf(data, width, left = 0, transect = "line")
}
\arguments{
\item{data}{as specified for \code{ds} and \code{ddf} (including a \code{size} column)}

\item{width}{right truncation}

\item{left}{left truncation (default 0, no left truncation)}

\item{transect}{\code{"line"} or \code{"point"} transect}
}
\description{
Create a detection function object for strip/plot surveys for use with
\code{dht2}.
}
\author{
David L Miller
}
