% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{ETP_Dolphin}
\alias{ETP_Dolphin}
\alias{ETP_Dolphin_units}
\title{Eastern Tropical Pacific spotted dolphin survey}
\format{A data frame with 1090 rows and 13 variables.
  \tabular{ll}{  
    \code{Region.Label} \tab stratum labels (only one)\cr
    \code{Area} \tab size (nmi) of each stratum \cr
    \code{Sample.Label} \tab transect labels \cr
    \code{Effort} \tab transect length (nmi) \cr
    \code{object} \tab object ID \cr
    \code{distance} \tab perpendicular distance (nmi) \cr
    \code{LnCluster} \tab natural log of cluster size \cr
    \code{Month} \tab month of detection \cr
    \code{Beauf.class} \tab Beaufort sea state \cr
    \code{Cue.type} \tab initial cue triggering detection \cr
    \code{Search.method} \tab observer method making the detection \cr
    \code{size} \tab cluster size \cr
    \code{Study.Area} \tab study area name \cr
  }}
\source{
Inter-American Tropical Tuna Commission
}
\description{
Observers aboard tuna vessels detecting dolphin schools along with a number
of possibly useful covariates for modelling the detection function.
}
\note{
Several different search methods included in these data
  \tabular{ll}{
  \code{0} \tab binoculars from crows nest \cr
  \code{2} \tab binoculars from elsewhere on ship \cr
  \code{3} \tab helicopter searching ahead of ship \cr
  \code{5} \tab radar detects of seabirds above dolphin schools \cr
  }
}
\seealso{
Several cue types were also recorded by observers.
  \tabular{ll}{
  \code{1} \tab seabirds above the school \cr
  \code{2} \tab water splashes \cr
  \code{3} \tab unspecified \cr
  \code{4} \tab floating objects such as logs \cr
  }
}
\keyword{datasets}
