% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{LTExercise}
\alias{LTExercise}
\alias{LTExercise_units}
\title{Simulated line transect survey data}
\format{
A \code{data.frame} with 106 rows and 7 variables
\itemize{
\item \code{Region.Label} strata names (single stratum)
\item \code{Area} size of study area (1 in this case, making abundance and density
equal)
\item \code{Sample.Label} transect ID
\item \code{Effort} length of transects (km)
\item \code{object} object ID
\item \code{distance} perpendicular distance (m)
\item \code{Study.Area} name of study area
}
}
\source{
Simulated data, from the distance sampling introductory course,
Centre for Research into Ecological & Environmental Modelling, University of
St Andrews.
}
\description{
Simulated line transect survey. Twelve transects, detection function is
half-normal.  True object density is 79.8 animals per km^2.
}
\note{
There is no unit object associated with this dataset
}
\keyword{datasets}
