% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_activity_fn.R
\name{make_activity_fn}
\alias{make_activity_fn}
\title{Multiplier bootstrap helper functions}
\usage{
make_activity_fn(..., detector_daily_duration = 24)
}
\arguments{
\item{\dots}{parameters specified by activity::fitact}

\item{detector_daily_duration}{by default we assume that detectors were able to detect animals for 24 hours, if they were only able to do this for some proportion of the day (say daylight hours), then adjust this argument accordingly}
}
\value{
a function which generates a single bootstrap estimate of
availability
}
\description{
Helper to use a models specified using \code{\link[activity:fitact]{activity::fitact}} to fit an
activity model and generate single realisations for bootstrapping with
\code{\link{bootdht}}.
}
\details{
Uses \code{\link[activity:fitact]{activity::fitact}} to generate single possible availability estimates
based on bootstraps. The function returns another function, which can be
passed to \code{bootdht}. It is recommended that you try out the function before
passing it to \code{\link{bootdht}}. See examples for a template for use.
}
\author{
David L Miller
}
