% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_dist_table.R
\docType{methods}
\name{p_dist_table}
\alias{p_dist_table}
\title{Distribution of probabilities of detection}
\arguments{
\item{object}{fitted detection function}

\item{bins}{how the results should be binned}

\item{proportion}{should proportions be returned as well as counts?}
}
\description{
Distribution of probabilities of detection
}
\note{
This function is located in the \code{mrds} package but the documentation
is provided here for easy access.
}
\examples{
\dontrun{
# example using a model for the minke data
data(minke)
# fit a model
result <- ds(minke, formula=~Region.Label)
# print table
p_dist_table(result)
# with proportions
p_dist_table(result, proportion=TRUE)
}
}
