% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dsmodel.R
\name{plot.dsmodel}
\alias{plot.dsmodel}
\title{Plot a fitted detection function}
\usage{
\method{plot}{dsmodel}(x, pl.den = 0, ...)
}
\arguments{
\item{x}{an object of class \code{dsmodel}.}

\item{pl.den}{shading density for histogram (default \code{0}, no shading)}

\item{...}{extra arguments to be passed to \code{\link[mrds:plot.ds]{plot.ds}}.}
}
\value{
\code{NULL}, just produces a plot.
}
\description{
This is just a simple wrapper around \code{\link[mrds:plot.ds]{plot.ds}}. See the
manual page for that function for more information.
}
\seealso{
\code{\link{add_df_covar_line}}
}
\author{
David L. Miller
}
