% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAIC.R
\name{QAIC}
\alias{QAIC}
\title{QIAC Information Criterion for detection functions when data is overdispersed}
\usage{
QAIC(object, ..., chat = NULL, k = 2)
}
\arguments{
\item{object}{a fitted detection function object}

\item{\dots}{additional fitted model objects.}

\item{chat}{a value of \deqn{\hat{c}}{chat} to be used in QIAC calculation}

\item{k}{penalty per parameter to be used; default 2}
}
\description{
Overdispersion causes AIC to select overly-complex models, so analysts
should specify the number/order of adjustment terms manually when fitting
distance sampling models to data from camera traps, rather than allowing
automated selection using AIC. Howe et al (2019) describes two methods for
performing model selection of distance sampling models in the face of
overdispersion. In both cases comparisons should be made between models
which use the same key function in their detection function specification.
}
\details{
The first method of Howe et al (2019) employs a two-step process. First, an
overdisersion factor (\deqn{\hat{c}_1}{chat1}) is computed for each key
function family from the most complex model in each family (derived from the
chi-squared goodness of fit test statistic divided by its degrees of
freedom). If multiple detection functions are supplied to \code{QAIC}, then
\code{chat} will be calculated automatically. Alternatively \code{chat} can be
supplied as an argument.

The second method (\deqn{\hat{c}_2}{chat2}) in the notation of Howe et al)
uses the number of distance observations recorded per independent encounter
between an animal and an observer. This quantity can be calculated from the
raw data. In camera trap surveys of solitary animals,
\deqn{\hat{c}_2}{chat2} would be the mean number of distance observations
recorded during a single pass by an animal in front of a trap. In surveys of
social animals employing human observers, \deqn{\hat{c}_2}{chat2} would be
the mean number of detected animals per detected group, and in camera trap
surveys of social animals \deqn{\hat{c}_2}{chat2} the mean number of
distance observations recorded during an encounter between a group of
animals and a CT. In this case one can calculate \deqn{\hat{c}_2}{chat2} and
use the \code{chat} argument to specify it for one or more models.
}
\examples{
\dontrun{
library(Distance)
data(minke)
model1 <- ds(minke, truncation=4, adjustment=NULL)
model2 <- ds(minke, truncation=4, adjustment="cos", order=2)
model4 <- ds(minke, truncation=4, adjustment="cos", order=c(2, 4))
model6 <- ds(minke, truncation=4, adjustment="cos", order=c(2, 4, 6))
# calculate QAIC
QAIC(model1, model2, model4, model6)

# using a pre-calculated chat
chat <- attr(QAIC(model1, model2, model4, model6), "chat")
QAIC(model1, chat=chat)
}
}
\references{
Howe, E. J., Buckland, S. T., Després-Einspenner, M.-L., & Kühl, H. S. (2019). Model selection with overdispersed distance sampling data. Methods in Ecology and Evolution, 10(1), 38–47. \doi{10.1111/2041-210X.13082}
}
\author{
David L Miller, based on code from Eric Rexstad and explanation from
Eric Howe.
}
