% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{Savannah_sparrow_1980}
\alias{Savannah_sparrow_1980}
\alias{Savannah_sparrow_1981}
\alias{Savannah_sparrow_1980_units}
\alias{Savannah_sparrow_1981_units}
\title{Savanna sparrow point transects}
\format{
\code{data.frame} with 468 observations (1980) and 448 observations
(1981) of 7 variables:
\itemize{
\item \code{Region.Label} stratum label (pasture ID)
\item \code{Area} stratum area (set to 1 so density is reported)
\item \code{Sample.Label} transect identifier
\item \code{Effort} number of visits
\item \code{object} object ID
\item \code{distance} radial distance (m)
\item \code{Study.Area} name of study area
}
}
\description{
Point transect data collected in Colorado 1980/81 to examine effect of
agricultural practices upon avian community.
}
\details{
Design consisted of point transects placed in multiple pastures (3 in 1980
and 4 in 1981). While many species were observed, only data for Savannah
sparrows (\emph{Passerculus sandwichensis}) are included here.

Data given here are different from the Distance for Windows example project.
Here each individual sighting is treated as an independent observation. This
corresponds to the analysis in Buckland et al. (2001) Section 8.7.  In the
Distance for Windows project objects are clusters of individuals. This
should not affect the results too greatly as most clusters were of size 1,
and so the results obtained should not be too far out.
}
\references{
Knopf, F.L., J.A. Sedgwick, and R.W. Cannon. (1988) Guild structure of a
riparian avifauna relative to seasonal cattle grazing.  The Journal of
Wildlife Management 52 (2): 280–290.  \doi{10.2307/3801235}
}
\keyword{datasets}
