% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdata.R
\docType{data}
\name{sikadeer}
\alias{sikadeer}
\alias{sikadeer_units}
\title{Sika deer pellet data from southern Scotland}
\format{
A \code{data.frame} with 1923 rows and 11 variables.
\itemize{
\item \code{Region.Label} stratum labels
\item \code{Area} size (ha) of each stratum
\item \code{Sample.Label} transect labels
\item \code{Defecation.rate} rate of dung production per individual per day
\item \code{Defecation.rate.SE} variability in defecation rate
\item \code{Decay.rate} time (days) for dung to become undetectable
\item \code{Decay.rate.SE} variability in decay rate
\item \code{Effort} transect length (km)
\item \code{object} object ID
\item \code{distance} perpendicular distance (cm)
\item \code{Study.Area} study area name
}
}
\description{
Because sika deer spend most of their time in woodland areas, abundance
estimates are based on pellet group counts.  Line transect methods were
applied to estimate deer pellet group density by geographic block.
}
\details{
Data presented here are from the Peebleshire portion of the study described
by Marques et al. (2001).
}
\references{
Marques, F.F.C., S.T. Buckland, D. Goffin, C.E. Dixon, D.L.  Borchers, B.A.
Mayle, and A.J. Peace. (2001). Estimating deer abundance from line transect
surveys of dung: sika deer in southern Scotland. Journal of Applied Ecology
38 (2): 349–363.  \doi{10.1046/j.1365-2664.2001.00584.x}
}
\keyword{datasets}
