% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ds_models.R
\name{summarize_ds_models}
\alias{summarize_ds_models}
\title{Make a table of summary statistics for detection function models}
\usage{
summarize_ds_models(..., sort = "AIC", output = "latex",
  delta_only = TRUE)
}
\arguments{
\item{...}{models to be summarised}

\item{sort}{column to sort by (default \code{"AIC"})}

\item{output}{should the output be given in \code{"latex"} compatible format or as \code{"plain"} text?}

\item{delta_only}{only output AIC differences (default \code{TRUE})}
}
\description{
Provide a summary table of useful information about fitted detection functions. This can be useful when paired with \code{knitr}s \code{kable} function.
}
\details{
Note that the column names are in LaTeX format, so if you plan to manipulate the resulting \code{data.frame} in R, you may wish to rename the columns for ease of access.
}
\author{
David L Miller
}

